+++ packages/geoview-core/public/templates/tests.html
@@ -120,7 +120,9 @@
       <div class="back-to-main">
         <a href="./index.html?tab=testing">← Back to Main</a>
         &nbsp;|&nbsp;
-        <a href="../../../../docs/app/testing/test-catalog.md" target="_blank">Test Catalog</a>
+        <a href="https://github.com/Canadian-Geospatial-Platform/geoview/blob/develop/docs/app/testing/test-catalog.md" target="_blank"
+          >Test Catalog</a
+        >
       </div>
     </div>
